server/requirements.txt
argon2-cffi==18.3.0
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector==2.1.6
mysql-connector-python==8.0.13
passlib==1.7.1
protobuf==3.6.1
pycparser==2.19
pymongo==3.7.2
pyOpenSSL==18.0.0
six==1.11.0
Werkzeug==0.14.1
